+++ tests/testthat/test-mutate.r
@@ -553,3 +553,8 @@ test_that("grouped mutate does not drop grouping attributes (#1020)", {
   a2 <- names(attributes(d %>% mutate(foo=1)))
   expect_equal( setdiff(a1, a2), character(0) )
 })
+
+test_that("grouped mutate errors on incompatible column type (#1641)", {
+  df <- data.frame(ID = rep(1:5, each = 3), x = 1:15) %>% group_by(ID)
+  expect_error( mutate(df, foo = mean), "unsupported type for column" )
+})