+++ src/matrixToDataFrame.cpp
@@ -1,16 +1,40 @@
 #include <Rcpp.h>
 using namespace Rcpp;
 
-// [[Rcpp::export]]
-List matrixToDataFrame(RObject x) {
-  SEXPTYPE type = TYPEOF(x);
-
+IntegerVector get_dim( const RObject& x){
   if (!x.hasAttribute("dim"))
     stop("`x` is not a matrix");
 
   IntegerVector dim = x.attr("dim");
   if (dim.size() != 2)
     stop("`x` is not a matrix");
+
+  return dim ;
+}
+
+CharacterVector get_names( const RObject& x, const IntegerVector& dim){
+  int nc = dim[1] ;
+  if( x.hasAttribute("dimnames") ){
+    List dimnames = x.attr("dimnames") ;
+    try {
+      CharacterVector res( dimnames[1] ) ;
+      return res ;
+    } catch(...){}
+  }
+
+  CharacterVector names( nc ) ;
+  for( int i=0; i<nc; i++){
+    names[i] = tfm::format( "V%d", (i+1) ) ;
+  }
+  return names ;
+}
+
+// [[Rcpp::export]]
+List matrixToDataFrame(RObject x) {
+  SEXPTYPE type = TYPEOF(x);
+
+  IntegerVector dim = get_dim(x) ;
+  CharacterVector names = get_names(x, dim) ;
 
   int nrow = dim[0], ncol = dim[1];
 
@@ -42,11 +66,7 @@ List matrixToDataFrame(RObject x) {
     }
   }
 
-  if (x.hasAttribute("dimnames")) {
-    List dimnames = x.attr("dimnames");
-    out.attr("names") = dimnames[1];
-  }
-
+  out.attr("names") = names ;
   out.attr("class") = CharacterVector::create("tbl_df", "tbl", "data.frame");
   out.attr("row.names") = IntegerVector::create(NA_INTEGER, -nrow);