+++ NEWS.md
@@ -1,6 +1,8 @@
 # dplyr 0.4.3.9000
 
-* `bind_rows` handles promotion to strings (#1538). 
+* `mutate` handles adding a factor that is all `NA` (#1645)
+
+* `bind_rows` handles promotion to strings (#1538).
 
 * `summarise` handles min/max of already summarised variable (#1622).